fix: resolve linting issues without eslint disable rules
