+++ apps/desktop/src/components/monaco.tsx
@@ -2,7 +2,7 @@ import type { RefObject } from 'react'
 import { noop } from '@conar/shared/utils/helpers'
 import { formatXml } from '@conar/shared/utils/xml'
 import { useMountedEffect } from '@conar/ui/hookas/use-mounted-effect'
-import { useTheme } from '@conar/ui/theme-observer'
+import { useTheme } from '@conar/ui/use-theme'
 import * as monaco from 'monaco-editor'
 import { vsPlusTheme } from 'monaco-sql-languages'
 import { useEffect, useEffectEvent, useRef } from 'react'
