+++ apps/desktop/src/components/table/table.tsx
@@ -1,7 +1,7 @@
 import type { ComponentProps } from 'react'
 import { ScrollArea } from '@conar/ui/components/custom/scroll-area'
 import { cn } from '@conar/ui/lib/utils'
-import { useTableContext } from './provider'
+import { useTableContext } from './use-table-context'
 
 export function Table({
   className,
