+++ apps/desktop/src/entities/database/components/cell-context.ts
@@ -1,0 +1,14 @@
+import type { UseMutateFunction } from '@tanstack/react-query'
+import type { Dispatch, SetStateAction } from 'react'
+import type { Column } from '../utils/table'
+import { createContext } from 'react'
+
+export const CellContext = createContext<{
+  value: string
+  setValue: Dispatch<SetStateAction<string>>
+  column: Column
+  initialValue: unknown
+  displayValue: string
+  update: UseMutateFunction<void, Error, { value: string | null, rowIndex: number }>
+  values?: string[]
+}>(null!)
