+++ apps/desktop/src/entities/database/components/table-cell-provider.tsx
@@ -1,23 +1,8 @@
-import type { UseMutateFunction } from '@tanstack/react-query'
-import type { Dispatch, SetStateAction } from 'react'
 import type { Column } from '../utils/table'
 import { useMutation } from '@tanstack/react-query'
-import { createContext, use, useMemo, useState } from 'react'
+import { useMemo, useState } from 'react'
 import { getEditableValue } from '../lib/render'
-
-const CellContext = createContext<{
-  value: string
-  setValue: Dispatch<SetStateAction<string>>
-  column: Column
-  initialValue: unknown
-  displayValue: string
-  update: UseMutateFunction<void, Error, { value: string | null, rowIndex: number }>
-  values?: string[]
-}>(null!)
-
-export function useCellContext() {
-  return use(CellContext)
-}
+import { CellContext } from './cell-context'
 
 export function TableCellProvider({
   children,
