+++ apps/desktop/src/entities/database/components/use-cell-context.ts
@@ -1,0 +1,6 @@
+import { use } from 'react'
+import { CellContext } from './cell-context'
+
+export function useCellContext() {
+  return use(CellContext)
+}
