+++ apps/desktop/src/lib/events.tsx
@@ -1,26 +1,6 @@
-import posthog from 'posthog-js'
 import { PostHogProvider } from 'posthog-js/react'
 
-export function initEvents() {
-  if (import.meta.env.DEV)
-    return null
-}
-
-interface IdentifyUserProps {
-  email: string
-  name: string
-}
-
-export function identifyUser(userId: string, properties: IdentifyUserProps): void
-export function identifyUser(userId: null, properties?: IdentifyUserProps): void
-export async function identifyUser(userId: string | null, properties?: IdentifyUserProps) {
-  if (userId) {
-    posthog.identify(userId, { ...properties, appVersion: await window.electron?.versions.app() })
-  }
-  else {
-    posthog.reset()
-  }
-}
+export { identifyUser, initEvents } from './events-utils'
 
 export function EventsProvider({ children }: { children: React.ReactNode }) {
   if (import.meta.env.DEV)
