+++ apps/desktop/src/routes/(protected)/-components/actions-center-store.ts
@@ -1,0 +1,9 @@
+import { Store } from '@tanstack/react-store'
+
+export const actionsCenterStore = new Store({
+  isOpen: false,
+})
+
+export function setIsOpen(isOpen: boolean) {
+  actionsCenterStore.setState(state => ({ ...state, isOpen }))
+}
