+++ packages/ui/src/components/field.tsx
@@ -209,8 +209,8 @@ function FieldError({
     return (
       <ul className="ml-4 flex list-disc flex-col gap-1">
         {uniqueErrors.map(
-          (error, index) =>
-            error?.message && <li key={index}>{error.message}</li>,
+          error =>
+            error?.message && <li key={error.message}>{error.message}</li>,
         )}
       </ul>
     )
