+++ packages/ui/src/components/toggle.tsx
@@ -2,30 +2,9 @@ import type { VariantProps } from 'class-variance-authority'
 import type * as React from 'react'
 import { cn } from '@conar/ui/lib/utils'
 import * as TogglePrimitive from '@radix-ui/react-toggle'
-import { cva } from 'class-variance-authority'
+import { toggleVariants } from './toggle.variants'
 
-const toggleVariants = cva(
-  'inline-flex items-center justify-center gap-2 rounded-md text-sm font-medium hover:bg-muted hover:text-muted-foreground disabled:pointer-events-none disabled:opacity-50 data-[state=on]:bg-accent data-[state=on]:text-accent-foreground [&_svg]:pointer-events-none [&_svg:not([class*=\'size-\'])]:size-4 [&_svg]:shrink-0 focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:ring-[3px] outline-none transition-[color,box-shadow] aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 aria-invalid:border-destructive whitespace-nowrap',
-  {
-    variants: {
-      variant: {
-        default: 'bg-transparent',
-        outline:
-          'border bg-transparent shadow-xs hover:bg-accent hover:text-accent-foreground',
-      },
-      size: {
-        default: 'h-9 px-2 min-w-9',
-        xs: 'h-6 px-1 min-w-6',
-        sm: 'h-8 px-1.5 min-w-8',
-        lg: 'h-10 px-2.5 min-w-10',
-      },
-    },
-    defaultVariants: {
-      variant: 'default',
-      size: 'default',
-    },
-  },
-)
+export { toggleVariants } from './toggle.variants'
 
 function Toggle({
   className,
@@ -43,4 +22,4 @@ function Toggle({
   )
 }
 
-export { Toggle, toggleVariants }
+export { Toggle }
