+++ packages/ui/src/use-theme.ts
@@ -1,0 +1,15 @@
+import type { ResolvedTheme, Theme } from './theme-store'
+import { useStore } from '@tanstack/react-store'
+import { getThemeStore, setTheme } from './theme-store'
+
+export function useTheme() {
+  const store = useStore(getThemeStore())
+
+  return {
+    theme: store.theme,
+    resolvedTheme: store.resolvedTheme,
+    setTheme,
+  }
+}
+
+export type { ResolvedTheme, Theme }
